Clean up UDP Packet buffers
